requirements.txt
mysql-connector
html5lib
bs4

